requirements.test.txt
pytest >= 3.6.0
pytest-cov >= 2.9.1
pytest-homeassistant-custom-component
